refactor registerFlow component
